restructured files for readability and maintainability
