+++ main.py
@@ -1,6 +1,6 @@
 import sys
 from PyQt6.QtWidgets import QApplication
-from GUI import MainWindow
+from MainWindow import MainWindow
 
 
 if __name__ == '__main__':
